refactor(notebook): extract NotebookSurface (modal|tile variants) (stage 7, PR 2)
Pull the entire Notebook body — file tree, live editor, context rail, fold
handles, tasks panel, and all load/save/send-to-chief logic — out of
NotebookBrowser into a reusable NotebookSurface that renders in two shapes:
- variant="modal": wrapped in the dialog shell (overlay + focus trap + header
  with the chief pulse + Close); NotebookBrowser owns this wrapper.
- variant="tile": a bare surface for an in-workspace tile (later PRs mount it).

The modal path is behaviorally byte-identical:
- NotebookBrowser stays the public component (same props/export) + re-exports
  parseNotebookHref/NotebookHref, so App, the harness, and the unit test are
  untouched.
- The surface stays MOUNTED while a modal is closed (state/refs survive a
  close→reopen), so the open/close gate became an `active` prop (= isOpen for
  the modal) rather than conditional mounting — all the isOpen-gated effects now
  read `active`.
- autoFold stays the hardcoded `false`; PR3 swaps it for a width-driven value.

New tile-only seams (inert for the modal, which passes neither): `initialPath`
(the file a tile reopens to) and `onOpenFile` (so a tile can persist the opened
path to its tile params).

No CHANGELOG: pure refactor, no user-visible change. Covered by the existing
NotebookBrowser unit suite (41) + the notebook-browser e2e harness (5), both green.
